Code review improvements
* Put factory in own class
* Remove direct references to TFramedTransport.Factory, so it's easier
  to see the new one is used everywhere
* Fix spelling of transport
* Fix formatting/comments in ThriftMaxFrameSizeIT
